tried to add subset to cellforestR to allow faceting on R (not working yet); moved R_FUNCTIONS to cellforestR to avoid passing path around; useconfig hook (won't always work yet, because branch process runs won't be cleared);
