+++ cellforest/__init__.py
@@ -1,12 +1,13 @@
 from pathlib import Path
 
-from dataforest import get_current_config, update_config
+from dataforest import get_current_config
 
 from cellforest.structures.counts.Counts import Counts
 from cellforest.templates.CellBranch import CellBranch
 from cellforest.templates.CellInterface import CellInterface
 from cellforest.templates.CellTree import CellTree
-
+from cellforest.utils.config.useconfig import useconfig
+from cellforest.utils.config.functions import CONFIG_OPTIONS, load_config, update_config
 
 from_sample_metadata = CellInterface.from_sample_metadata
 from_input_dirs = CellInterface.from_input_dirs
