+++ cellforest/processes/__init__.py
@@ -1,3 +1,3 @@
-from pathlib import Path
-
-R_FUNCTIONS_FILEPATH = Path(__file__).parent / "scripts/functions.R"
+# from pathlib import Path
+#
+# R_FUNCTIONS_FILEPATH = Path(__file__).parent / "scripts/functions.R"
