+++ cellforest/processes/processes/cluster/process.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from dataforest.hooks import dataprocess
 
-from cellforest.processes import R_FUNCTIONS_FILEPATH
+# from cellforest.processes import R_FUNCTIONS_FILEPATH
 from cellforest.utils.r.run_r_script import run_r_script_logged
 
 R_CLUSTER_SCRIPT = Path(__file__).parent / "cluster.R"
@@ -27,6 +27,6 @@ def cluster(branch: "CellBranch", run_name: str):
         params["num_pcs"],
         params["res"],
         params["eps"],
-        R_FUNCTIONS_FILEPATH,
+        # R_FUNCTIONS_FILEPATH,
     ]
     run_r_script_logged(branch, R_CLUSTER_SCRIPT, arg_list, run_name)
