+++ cellforest/processes/processes/normalize/sctransform.R
@@ -7,14 +7,25 @@ min_genes <- as.numeric(args[4])
 max_genes <- as.numeric(args[5])
 min_cells <- as.numeric(args[6])
 perc_mito_cutoff <- as.numeric(args[7])
-r_functions_filepath <- args[8]
+#r_functions_filepath <- args[8]
 
-output_corrected_umi_path <- args[9]
-output_pearson_residual_path <- args[10]
+output_corrected_umi_path <- args[8]
+output_pearson_residual_path <- args[9]
 
-source(r_functions_filepath)
+#source(r_functions_filepath)
 
-seurat_object <- create_seurat_object(input_tenx_directory_path, input_metadata_path, min_cells)
+run_sctransform <- function(seurat_object, corrected_umi_output_path, pearson_residual_output_path) {
+  print("sctransform")
+  print(date())
+  seurat_object = SCTransform(object = seurat_object, verbose = TRUE)
+
+  writeMM(seurat_object$SCT@counts, file = corrected_umi_output_path)
+  write.table(seurat_object$SCT@scale.data, file = pearson_residual_output_path, sep = "\t", quote = FALSE)
+
+  return(seurat_object)
+}
+
+seurat_object <- srat_from_tenx(input_tenx_directory_path, input_metadata_path, min_cells)
 print("reading metadata"); print(date())
 metadata <- read_tsv(input_metadata_path)
 print("metadata filter"); print(date())
