+++ cellforestR/NAMESPACE
@@ -1,12 +1,28 @@
 # Generated by roxygen2: do not edit by hand
 
+export(add_genes_perc_meta)
 export(cellforest_load)
 export(example_spec_py)
 export(example_spec_r)
+export(filter_cells)
+export(find_clusters)
 export(get_seurat_object)
+export(metadata_filter_objs)
+export(metadata_filter_paths)
+export(run_pca)
+export(save_embeddings)
+export(srat_from_tenx)
 importFrom(Seurat,AddMetaData)
 importFrom(Seurat,CreateDimReducObject)
+importFrom(Seurat,CreateSeuratObject)
 importFrom(Seurat,DefaultAssay)
+importFrom(Seurat,FetchData)
+importFrom(Seurat,FindClusters)
+importFrom(Seurat,FindNeighbors)
+importFrom(Seurat,Idents)
+importFrom(Seurat,PercentageFeatureSet)
+importFrom(Seurat,Read10X)
+importFrom(Seurat,RunPCA)
 importFrom(glue,glue)
 importFrom(reticulate,import)
 importFrom(reticulate,py)
