+++ tests/test_quick.py
@@ -4,5 +4,6 @@ from tests.test_process import *
 from tests.test_data_ops import *
 from tests.test_historical import *
 from tests.test_version_control import *
-from tests.test_cellforestR import test_cellforest_load_last_process
+
+# from tests.test_cellforestR import test_cellforest_load_last_process
 from tests.test_qc_plots import test_root_plots, test_ggplot2_plots
